🎈 perf: Refactor singleton using weakref so that non-ref instance can be released by gc
